% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_cluster_cov.R
\name{cluster_toeplitz}
\alias{cluster_toeplitz}
\title{#create single cluster type covariance matrix with fixed sparsity.
#s is number of non zero components in each column of covariance}
\usage{
cluster_toeplitz(p, rho = 0.75, root = 2, s = 6)
}
\description{
#create single cluster type covariance matrix with fixed sparsity.
#s is number of non zero components in each column of covariance
}
