% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot.R
\name{forestplot}
\alias{forestplot}
\title{draw forest plot}
\usage{
forestplot(
  datain,
  XVar,
  XLowVar,
  XUppVar,
  YOrderVar,
  XLine,
  YRefVar,
  YRefWidth,
  Anno1Var = NULL,
  Anno1VarX = NULL,
  Anno1VarLabel = NULL,
  Anno2Var = NULL,
  Anno2VarX = NULL,
  Anno2VarLabel = NULL,
  Anno3Var = NULL,
  Anno3VarX = NULL,
  Anno3VarLabel = NULL,
  Anno4Var = NULL,
  Anno4VarX = NULL,
  Anno4VarLabel = NULL,
  Anno5Var = NULL,
  Anno5VarX = NULL,
  Anno5VarLabel = NULL,
  Anno6Var = NULL,
  Anno6VarX = NULL,
  Anno6VarLabel = NULL,
  AnnoLineLabel,
  XaxisValue,
  XaxisLabel,
  XaxisType,
  leftmargin,
  rightmargin,
  figwidth,
  figheight,
  FigureName = "forestplot"
)
}
\arguments{
\item{datain}{a dataframe or tibble}

\item{XVar}{X axis column}

\item{XLowVar}{lower limit of X axis column}

\item{XUppVar}{upper limit of X axis column}

\item{YOrderVar}{Y axis column (numeric: from 1 to N)}

\item{XLine}{reference line in X axis}

\item{YRefVar}{add background color in Y axis column}

\item{YRefWidth}{background color width}

\item{Anno1Var}{first annotation column}

\item{Anno1VarX}{X axis of first annotation column}

\item{Anno1VarLabel}{label of first annotation column}

\item{Anno2Var}{2nd annotation column}

\item{Anno2VarX}{X axis of 2nd annotation column}

\item{Anno2VarLabel}{label of 2nd annotation column}

\item{Anno3Var}{3rd annotation column}

\item{Anno3VarX}{X axis of 3rd annotation column}

\item{Anno3VarLabel}{label of 3rd annotation column}

\item{Anno4Var}{4th annotation column}

\item{Anno4VarX}{X axis of 4th annotation column}

\item{Anno4VarLabel}{label of 4th annotation column}

\item{Anno5Var}{5th annotation column}

\item{Anno5VarX}{X axis of 5th annotation column}

\item{Anno5VarLabel}{label of 5th annotation column}

\item{Anno6Var}{6th annotation column}

\item{Anno6VarX}{X axis of 6th annotation column}

\item{Anno6VarLabel}{label of 6th annotation column}

\item{AnnoLineLabel}{label of reference line annotation column}

\item{XaxisValue}{axis value displayed in X axis}

\item{XaxisLabel}{axis label displayed in X axis}

\item{XaxisType}{X axis type}

\item{leftmargin}{reserve area in the left of figure}

\item{rightmargin}{reserve area in the right of figure}

\item{figwidth}{figure width}

\item{figheight}{figure height}

\item{FigureName}{figure name}
}
\value{
a figure as ggplot2 object
}
\description{
draw forest plot
}
\examples{
forestplot(
datain = final,
XVar = final$means,
XLowVar = final$yerrl,
XUppVar = final$yerru,
YOrderVar = final$yord,
XLine = 1,
YRefVar = final$ref,
YRefWidth = 4.9,
Anno1Var = final$col1,
Anno1VarX = 10^-3,
Anno1VarLabel = "D14 Subgroup",
Anno2Var = final$col2,
Anno2VarX = 3*10^-2,
Anno2VarLabel = "V01E GMT",
Anno3Var = final$col3,
Anno3VarX = 10^-1,
Anno3VarLabel = "V01E-2 GMT",
Anno4Var = final$col5,
Anno4VarX = 4*10^1,
Anno4VarLabel = "LS GMR 95\% CI",
Anno5Var = NULL,
Anno5VarX = NULL,
Anno5VarLabel = NULL,
Anno6Var = NULL,
Anno6VarX = NULL,
Anno6VarLabel = NULL,
AnnoLineLabel = "LS GMR 95\% CI",
XaxisValue = c(0.1, 1, 10),
XaxisLabel = c("<0.1", "1", ">=10"),
XaxisType = "log",
leftmargin = 9.5,
rightmargin = 4.5,
figwidth = 9,
figheight = 4,
FigureName = "forestplot"
)
}
