% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmtplot.R
\name{gmtplot}
\alias{gmtplot}
\title{draw GMT plot}
\usage{
gmtplot(datain, FigureName)
}
\arguments{
\item{datain}{a dataframe or tibble}

\item{FigureName}{a character}
}
\value{
a figure as ggplot2 object
}
\description{
draw GMT plot
}
\examples{
gmtplot(adis,"gmt_plot")
}
