% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmplot.R
\name{kmplot}
\alias{kmplot}
\title{Draw KM plot}
\usage{
kmplot(
  datain,
  GrpVar = TRTAN,
  GrpLabel,
  AVALVar = AVAL,
  CNSRVar = CNSR,
  ByTime,
  XLabel,
  YLabel,
  RiskLabel,
  colorSet,
  ShowAreaYN = FALSE,
  LineMedianYN = FALSE,
  PvalYN = FALSE,
  FigureName = "kmplot"
)
}
\arguments{
\item{datain}{a dataframe or tibble}

\item{GrpVar}{group by numeric Column}

\item{GrpLabel}{a vector for group by column label}

\item{AVALVar}{analysis time numeric column}

\item{CNSRVar}{censor numeric column}

\item{ByTime}{by step in X axis}

\item{XLabel}{a character for X axis label}

\item{YLabel}{a character for Y axis label}

\item{RiskLabel}{a character for risktable label}

\item{colorSet}{a vector for color}

\item{ShowAreaYN}{whether draw area plot}

\item{LineMedianYN}{whether draw reference line when y = 0.5}

\item{PvalYN}{whether display P value}

\item{FigureName}{a character for figure name}
}
\value{
a figure as ggplot2 object
}
\description{
Draw KM plot
}
\examples{
kmplot(datain = adtte,
GrpVar = adtte$COHORTN,
GrpLabel = c("Group A", "B","C"),
AVALVar = adtte$AVAL,
CNSRVar = adtte$CNSR,
ByTime = 2,
XLabel = "XXXX",
YLabel = "YYYY",
RiskLabel = "Number at risk",
colorSet = c("grey", "blue", "red"),
FigureName = "kmplot")

}
