% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_eval.R
\name{ls_eval}
\alias{ls_eval}
\title{Least squares evaluation}
\usage{
ls_eval(x, tau, type = "mean", param_pre, param_post = NULL)
}
\arguments{
\item{x}{an n by p matrix of observations.}

\item{tau}{a real parameter between (0,1) for point evaluation or 'all'.}

\item{mat_pre}{an p by p-1 of coefficients for the second segment}
}
\value{
\item{}{A real value if \code{tau} is in (0,1)}
\item{}{A vector of reals if \code{tau="all"}}
}
\description{
Computes the least squares cost of a data matrix at a changepoint
or for all values of the changepoint, given matrices of parameters.
}
\details{
If \code{tau="all"} the evaluation happens for all values of the
changepoint and a seqeunce of real values is returned. Otherwise if \code{tau}
is in (0,1) a single value at that point is returned.
}
\examples{
p = 50; n = 225; t0 = 0.3
params = gen_param(p = p, reg_param = T)
sig1 = params$pre_cov; sig2 = params$post_cov
x = gen_data(n = n, p = p, t0 = t0, sig1, sig2)$data
pre_mat = params$mu_mat
post_mat = params$gamma_mat
# Evaluation at a point
ls_eval(x, 0.5, type = "graph", pre_mat, post_mat)
# Evaluation everywhere
ls_eval(x, "all",  type = "graph", pre_mat, post_mat)
}
\author{
Konstantinos Tsampourakis <kostas.tsampourakis@gmail.com>
}
