% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_mat.R
\name{pred_mat}
\alias{pred_mat}
\title{Prediction matrix calculation.}
\usage{
pred_mat(x, coef_mat)
}
\arguments{
\item{x}{an n by p matrix of observations.}

\item{coef_mat}{an p by p-1 of coefficients of the linear model
for each component.}
}
\value{
\item{pred}{A n by p matrix of the predicted values.}
}
\description{
Computes the matrix of the predicted values according to a linear
model for all elements of the data matrix.
}
\examples{
n = 300; p = 100
x = matrix(rnorm(n * p), n, p)
mod = coef_estimate(x)
mod = pred_mat(x, coef.mat = mod$coef.mat)
pred = mod$predicted
}
