% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spider.R
\name{spiderplot}
\alias{spiderplot}
\title{Draw spiderplot}
\usage{
spiderplot(
  datain,
  GrpSubjVar,
  XaxisVar,
  YaxisVar,
  GrpVar,
  YLine1 = 20,
  Yline2 = -30,
  colorSet = c("blue", "red", "green", "black", "yellow"),
  XLabel,
  YLabel,
  FigureName = "spiderplot"
)
}
\arguments{
\item{datain}{a dataframe or tibble}

\item{GrpSubjVar}{group by column for each line}

\item{XaxisVar}{X axis column}

\item{YaxisVar}{Y axis column}

\item{GrpVar}{group by Column for line or point}

\item{YLine1}{1st reference line in Y axis}

\item{Yline2}{2nd reference line in Y axis}

\item{colorSet}{a vector for color}

\item{XLabel}{a character for X axis label}

\item{YLabel}{a character for Y axis label}

\item{FigureName}{a character for figure name}
}
\value{
a figure as ggplot2 object
}
\description{
Draw spiderplot
}
\examples{
spiderplot(datain = spiderdata,
GrpSubjVar = spiderdata$Subject,
XaxisVar = spiderdata$durx,
YaxisVar = spiderdata$chgx,
GrpVar = spiderdata$BOR,
YLine1 = 20,
Yline2 = -30,
colorSet = c("blue", "red", "green", "black", "yellow"),
XLabel = "Time Since Treatment Initiation (months)",
YLabel = "Tumor Change From Baseline",
FigureName = "spiderplot")
}
