% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_and_fit.R
\name{split_and_fit}
\alias{split_and_fit}
\title{Split and fit function.}
\usage{
split_and_fit(
  x,
  t_split,
  type = "mean",
  lambda_vec = NULL,
  cv = 5,
  NumCores = 4,
  par = F,
  length_lam = 50,
  lam_sep = 1.75,
  l_end = 2,
  k_seq = seq(0, 0.5, length.out = 50)
)
}
\arguments{
\item{x}{an n by p matrix of observations.}

\item{t_split}{a split value between 0 and 1.}

\item{type}{a string of characters denoting the detection type : "mean" or
"graph"}

\item{lambda_vec}{a grid of values of the regularization parameter
to be used for cross-validation.}

\item{cv}{number of cross-validation folds.}

\item{NumCores}{number of processor cores to be used}

\item{par}{If TRUE parallelization of the lasso estimation is performed.}

\item{length_lam}{length of grid.}

\item{lam_sep}{numeric value. Used in the regularizer grid if not provided by
the user.}

\item{l_end}{end value of the lambda sequence}

\item{bic_const}{penalty used in the BIC evaluation.}

\item{start_grid}{grip of values for the initial split location}
}
\value{
\item{param_pre}{Parameters for the first segment}
\item{param_post}{Parameters for the second segment}
}
\description{
Split the data at a given point and fit to obtain parameters
for each segment.
}
\details{
This function separates the data into segments at a point
\code{n*t_split} and estimates parameters. If  \code{type="mean"} the
parameters are the soft-thresholded sample means of the segments.
If \code{type="graph"} the parameters are estimated using the procedure of
\insertCite{yuan2010high}{cpgraph}. In this case the parameters are obtained
using the \code{glmnet} regularization path with parameters specified in
\code{lambda_vec}. If \code{lambda_vec} is omitted the path is set to
\deqn{seq(lam_sep*(sqrt(log(max(p,n))/n),l_end,length.out=length_lam)}
If \code{par=TRUE} the coefficient estimation is done in parallel using
the library \code{parallel} with the number of CPU cores set to \code{NumCores}.
}
\examples{
p=50;n=225;t0=0.2
params=gen.param(p=p, reg_param = TRUE)
sig1=params$pre.cov; sig2=params$post.cov
x=gen.data(n=n,p=p,t0=t0,sig1,sig2)$data
# split in the middle of the dataset
snf = split_and_fit(x, t_split = 0.5)
pre = snf$pre_mat
post = snf$post_mat
}
\references{
\insertAllCited
}
\author{
Konstantinos Tsampourakis <kostas.tsampourakis@gmail.com>
}
