% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swimmerplot.R
\name{swimmerplot}
\alias{swimmerplot}
\title{Draw swimmer plot}
\usage{
swimmerplot(
  datain,
  YaxisVar,
  BarVar = NULL,
  BarColor = NULL,
  BarWidth = NULL,
  BarLabel = NULL,
  BarLabelY = NULL,
  Line1VarMin = NULL,
  Line1VarMax = NULL,
  Line1color = NULL,
  Line1Width = NULL,
  Line1Label = NULL,
  Line1LabelY = NULL,
  Line2VarMin = NULL,
  Line2VarMax = NULL,
  Line2color = NULL,
  Line2Width = NULL,
  Line2Label = NULL,
  Line2LabelY = NULL,
  Line3VarMin = NULL,
  Line3VarMax = NULL,
  Line3color = NULL,
  Line3Width = NULL,
  Line3Label = NULL,
  Line3LabelY = NULL,
  Point1Var = NULL,
  Point1Color = NULL,
  Point1Shape = NULL,
  Point1Label = NULL,
  Point1LabelY = NULL,
  Point2Var = NULL,
  Point2Color = NULL,
  Point2Shape = NULL,
  Point2Label = NULL,
  Point2LabelY = NULL,
  Point3Var = NULL,
  Point3Color = NULL,
  Point3Shape = NULL,
  Point3Label = NULL,
  Point3LabelY = NULL,
  Point4Var = NULL,
  Point4Color = NULL,
  Point4Shape = NULL,
  Point4Label = NULL,
  Point4LabelY = NULL,
  Point5Var = NULL,
  Point5Color = NULL,
  Point5Shape = NULL,
  Point5Label = NULL,
  Point5LabelY = NULL,
  Point6Var = NULL,
  Point6Color = NULL,
  Point6Shape = NULL,
  Point6Label = NULL,
  Point6LabelY = NULL,
  Point7Var = NULL,
  Point7Color = NULL,
  Point7Shape = NULL,
  Point7Label = NULL,
  Point7LabelY = NULL,
  Point8Var = NULL,
  Point8Color = NULL,
  Point8Shape = NULL,
  Point8Label = NULL,
  Point8LabelY = NULL,
  Segment1Var = NULL,
  Segment1Color = NULL,
  Segment1Label = NULL,
  Segment1LabelY = NULL,
  XaxisLabel,
  YaxisLabel,
  figwidth,
  figheight,
  FigureName = "swimmerplot"
)
}
\arguments{
\item{datain}{a dataframe or tibble}

\item{YaxisVar}{Y axis column}

\item{BarVar}{barplot column}

\item{BarColor}{barplot color}

\item{BarWidth}{bar width}

\item{BarLabel}{bar label in legend}

\item{BarLabelY}{position of bar label in Y axis}

\item{Line1VarMin}{1st line X axis min column}

\item{Line1VarMax}{1st line X axis max column}

\item{Line1color}{1st line color}

\item{Line1Width}{1st line width}

\item{Line1Label}{1st line label}

\item{Line1LabelY}{position of 1st line label in Y axis}

\item{Line2VarMin}{2nd line X axis min column}

\item{Line2VarMax}{2nd line X axis max column}

\item{Line2color}{2nd line color}

\item{Line2Width}{2nd line width}

\item{Line2Label}{2nd line label}

\item{Line2LabelY}{position of 2nd line label in Y axis}

\item{Line3VarMin}{3rd line X axis min column}

\item{Line3VarMax}{3rd line X axis max column}

\item{Line3color}{3rd line color}

\item{Line3Width}{3rd line width}

\item{Line3Label}{3rd line label}

\item{Line3LabelY}{position of 3rd line label in Y axis}

\item{Point1Var}{1st point column}

\item{Point1Color}{1st point color}

\item{Point1Shape}{1st point shape}

\item{Point1Label}{1st point label}

\item{Point1LabelY}{1st point label position in Y axis}

\item{Point2Var}{2nd point column}

\item{Point2Color}{2nd point color}

\item{Point2Shape}{2nd point shape}

\item{Point2Label}{2nd point label}

\item{Point2LabelY}{2nd point label position in Y axis}

\item{Point3Var}{3rd point column}

\item{Point3Color}{3rd point color}

\item{Point3Shape}{3rd point shape}

\item{Point3Label}{3rd point label}

\item{Point3LabelY}{3rd point label position in Y axis}

\item{Point4Var}{4th point column}

\item{Point4Color}{4th point color}

\item{Point4Shape}{4th point shape}

\item{Point4Label}{4th point label}

\item{Point4LabelY}{4th point label position in Y axis}

\item{Point5Var}{5th point column}

\item{Point5Color}{5th point color}

\item{Point5Shape}{5th point shape}

\item{Point5Label}{5th point label}

\item{Point5LabelY}{5th point label position in Y axis}

\item{Point6Var}{6th point column}

\item{Point6Color}{6th point color}

\item{Point6Shape}{6th point shape}

\item{Point6Label}{6th point label}

\item{Point6LabelY}{6th point label position in Y axis}

\item{Point7Var}{7th point column}

\item{Point7Color}{7th point color}

\item{Point7Shape}{7th point shape}

\item{Point7Label}{7th point label}

\item{Point7LabelY}{7th point label position in Y axis}

\item{Point8Var}{8th point column}

\item{Point8Color}{8th point color}

\item{Point8Shape}{8th point shape}

\item{Point8Label}{8th point label}

\item{Point8LabelY}{8th point label position in Y axis}

\item{Segment1Var}{1st segment column}

\item{Segment1Color}{1st segment color}

\item{Segment1Label}{1st segment label}

\item{Segment1LabelY}{1st segment label position in Y axis}

\item{XaxisLabel}{X axis label}

\item{YaxisLabel}{Y axis label}

\item{figwidth}{figure width}

\item{figheight}{figure height}

\item{FigureName}{figure name}
}
\value{
a figure as ggplot2 object
}
\description{
Draw swimmer plot
}
\examples{
swimmerplot(datain = swimmerdata,
YaxisVar = swimmerdata$SUBJID,
BarVar = swimmerdata$trtedy,
BarColor = "grey",
BarWidth = 0.01,
BarLabel = "研究持续时间",
BarLabelY = 17,
Line1VarMin = swimmerdata$startdr,
Line1VarMax = swimmerdata$enddy,
Line1color = "blue",
Line1Width = 2,
Line1Label = "缓解持续时间",
Line1LabelY = 15,
Point1Var = swimmerdata$pddy,
Point1Color = "black",
Point1Shape = 15,
Point1Label = "疾病进展",
Point1LabelY = 13,
Point2Var = swimmerdata$dthdy,
Point2Color = "red",
Point2Shape = 16,
Point2Label = "死亡",
Point2LabelY = 11,
Segment1Var = swimmerdata$exongo,
Segment1Color = "brown",
Segment1Label = "治疗持续",
Segment1LabelY = 9,
XaxisLabel = "相对首次给药时间月",
YaxisLabel = "受试者编号",
figwidth = 9,
figheight = 5,
FigureName = "swimmerplot")
}
