% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfallplot.R
\name{waterfallplot}
\alias{waterfallplot}
\title{Draw waterfall plot}
\usage{
waterfallplot(
  datain,
  BORVar,
  CHGVar,
  PercentLine1 = FALSE,
  PercentLine2 = FALSE,
  YLabel,
  colorSet = c("green", "blue", "yellow", "red", "grey"),
  FigureName = "waterfallplot"
)
}
\arguments{
\item{datain}{a dataframe or tibble}

\item{BORVar}{BOR column in dataframe}

\item{CHGVar}{percent column in Y axis}

\item{PercentLine1}{reference line 1}

\item{PercentLine2}{reference line 2}

\item{YLabel}{a character for Y axis label}

\item{colorSet}{a vector for color}

\item{FigureName}{a character for figure name}
}
\value{
a figure as ggplot2 object
}
\description{
Draw waterfall plot
}
\examples{
waterfallplot(water,
BORVar = water$BOR,
CHGVar = water$chgx,
PercentLine1 = FALSE,
PercentLine2 = FALSE,
YLabel = "Change from Baseline",
colorSet = c("green", "blue", "yellow", "red", "grey"),
FigureName = "waterfallplot")

}
