% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aetoxgrplot.R
\name{aetoxgrplot}
\alias{aetoxgrplot}
\title{Draw AETOXGR plot}
\usage{
aetoxgrplot(
  dataADSL,
  dataADAE,
  GrpVar = TRTAN,
  GrpADSLVar = TRT01AN,
  GrpLabel,
  ATOXGRNVar = ATOXGRN,
  AEDECODVar = AEDECOD,
  Ylabel = c("不良事件发生率（\%）"),
  LegendLabel = c("一级", "二级", "三级及以上"),
  colorSet = c("grey", "blue", "red"),
  FigureName = "aetoxgrplot"
)
}
\arguments{
\item{dataADSL}{a dataframe or tibble for population}

\item{dataADAE}{a dataframe or tibble for events}

\item{GrpVar}{group by numeric Column for events}

\item{GrpADSLVar}{group by numeric Column for population}

\item{GrpLabel}{a vector for group by column label}

\item{ATOXGRNVar}{AETOXGR numeric column}

\item{AEDECODVar}{AEDECOD coding}

\item{Ylabel}{a character for Y axis label}

\item{LegendLabel}{a vector for legend label}

\item{colorSet}{a vector for color}

\item{FigureName}{a character for figure name}
}
\value{
a figure as ggplot2 object
}
\description{
Draw AETOXGR plot
}
\examples{
aetoxgrplot(dataADSL = adsl,
dataADAE = adae,
GrpVar = TRTAN,
GrpADSLVar = TRT01AN,
GrpLabel = c("V01A", "V01B"),
ATOXGRN = ATOXGRN,
AEDECOD = AEDECOD,
Ylabel = c("不良事件发生率（\%）"),
colorSet = c("red", "blue", "grey"),
FigureName = "aetox")
}
