% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmtplot.R
\name{gmtplot}
\alias{gmtplot}
\title{draw GMT plot}
\usage{
gmtplot(
  datain,
  GrpVar,
  AvisintVar,
  Aval,
  Base,
  YLabel,
  LegendLabel,
  FigureName
)
}
\arguments{
\item{datain}{a dataframe or tibble}

\item{GrpVar}{group by Column}

\item{AvisintVar}{analysis visit numeric column}

\item{Aval}{analysis value numeric column}

\item{Base}{baseline column}

\item{YLabel}{a character for Y axis label}

\item{LegendLabel}{a vector for Legend label}

\item{FigureName}{a character for figure name}
}
\value{
a figure as ggplot2 object
}
\description{
draw GMT plot
}
\examples{
gmtplot(adis,"gmt_plot")
}
