% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmtplot.R
\name{gmtplot}
\alias{gmtplot}
\title{draw GMT plot}
\usage{
gmtplot(
  datain,
  GrpVar = TRTAN,
  GrpLabel,
  AvisitnVar = AVISITN,
  AvisintVal,
  AvisitLabel,
  Aval = AVAL,
  Base = BASE,
  YLabel,
  LegendLabel,
  colorSet,
  LineYN = FALSE,
  LegendYN = FALSE,
  FigureName = "gmtplot"
)
}
\arguments{
\item{datain}{a dataframe or tibble}

\item{GrpVar}{group by numeric Column}

\item{GrpLabel}{a vector for group by column label}

\item{AvisitnVar}{analysis visit numeric column}

\item{AvisintVal}{analysis visit value}

\item{AvisitLabel}{a vector for analysis visit label}

\item{Aval}{analysis value numeric column}

\item{Base}{baseline column}

\item{YLabel}{a character for Y axis label}

\item{LegendLabel}{a vector for Legend label}

\item{colorSet}{a vector for color}

\item{LineYN}{whether draw geom_line}

\item{LegendYN}{whether draw legend}

\item{FigureName}{a character for figure name}
}
\value{
a figure as ggplot2 object
}
\description{
draw GMT plot
}
\examples{
gmtplot(adis,"gmt_plot")
}
